apps/memory-integrity-engine/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
numpy>=1.26
pydantic>=2.7
httpx>=0.27
qdrant-client>=1.12
grpcio>=1.80
protobuf>=5.26
scikit-learn>=1.4
psycopg[binary]>=3.1
prometheus-client>=0.20
nats-py>=2.6
# Optional: set MEMORY_SENTENCE_TRANSFORMER_MODEL=all-MiniLM-L6-v2 (adds ~2GB torch wheels)
# sentence-transformers>=3.0
